refactor: 이미지 업로드 커스텀 훅 및 컴포넌트 적용
