+++ src/app/(route)/dogstagram/writing/_component/DogStagramWriting.tsx
@@ -12,7 +12,7 @@ import InputImage from "@/app/(commons)/_component/imageupload/InputImage";
 export default function DogStargramWriting() {
   const {
     register,
-    formState: { errors, isValid = false },
+    formState: { errors, isValid },
     watch,
   } = useForm<ProfileFormData>({
     criteriaMode: "all",
